Autoload Types Moved build_coercer in types.rb
